optimized input form
