+++ src/components/Footer/Footer.module.css
@@ -3,7 +3,7 @@
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
-  position: absolute;
+  position: absolute !important;
   bottom: 0;
   left: 0;
   justify-content: space-between;
