+++ src/components/SingleDayData/SingleDayData.module.css
@@ -3,4 +3,5 @@
   display: flex;
   flex-wrap: wrap;
   margin: auto;
+  justify-content: space-between;
 }
