+++ src/containers/App/App.css
@@ -1,7 +1,7 @@
 .App {
   text-align: center;
   height: 650px;
-  max-width: 400px;
+  width: 400px;
   font-size: 8px;
   background-color: #e0e0de;
   margin: auto;
